pyproject.toml
[project]
name = "andersen-tech-task"
version = "0.1.0"
description = ""
authors = [
    { name = "VoiceOfDarkness", email = "abil.samedov2016@yandex.ru" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "dependency-injector (>=4.46.0,<5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.11.6,<0.12.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"